+++ src/StateSmith.Cli/CHANGELOG.md
@@ -8,7 +8,7 @@ Test/interim releases are not documented here.
 
 ---
 
-## wip
+## [0.9.2]
 ### Change (minor)
 - `run` command for "diagram-only" projects now prints the proper path for created files.
     - https://github.com/StateSmith/StateSmith/issues/345